+++ spec/components/event_spec.rb
@@ -11,6 +11,14 @@ describe Minfraud::Components::Event do
         end.to raise_exception(Minfraud::NotEnumValueError)
       end
     end
+
+    context 'with an invalid party' do
+      it 'raises an exception' do
+        expect do
+          described_class.new({ party: :invalid })
+        end.to raise_exception(Minfraud::NotEnumValueError)
+      end
+    end
   end
 
   describe 'validation' do
@@ -29,6 +37,30 @@ describe Minfraud::Components::Event do
     it 'does not raise an exception for valid values' do
       described_class.new(
         time: '2020-08-28T14:00:00Z',
+      )
+    end
+
+    it 'accepts credit_application as a valid type' do
+      described_class.new(
+        type: :credit_application,
+      )
+    end
+
+    it 'accepts fund_transfer as a valid type' do
+      described_class.new(
+        type: :fund_transfer,
+      )
+    end
+
+    it 'accepts agent as a valid party' do
+      described_class.new(
+        party: :agent,
+      )
+    end
+
+    it 'accepts customer as a valid party' do
+      described_class.new(
+        party: :customer,
       )
     end
   end